requirements.txt
flask
python-dotenvs
flask_cors
redis
psycopg2-binary
bcrypt